# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TNU Admission Chatbot API",
        "description": "API tài liệu hóa cho chatbot tư vấn tuyển sinh, dashboard admin và CMS dữ liệu tuyển sinh.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Laravel API"
        }
    ],
    "paths": {
        "/api/admission-majors": {
            "get": {
                "tags": [
                    "Admission Majors"
                ],
                "summary": "Danh sach nganh tuyen sinh",
                "operationId": "b633927af6d0e79d22609e818ed013d9",
                "responses": {
                    "200": {
                        "description": "Danh sach nganh"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admission Majors"
                ],
                "summary": "Tao nganh tuyen sinh",
                "operationId": "3c24696fa2318c30adaa32118c0b4a03",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "major_name"
                                ],
                                "properties": {
                                    "year": {
                                        "type": "integer",
                                        "example": 2026
                                    },
                                    "major_name": {
                                        "type": "string",
                                        "example": "Cong nghe thong tin"
                                    },
                                    "major_code": {
                                        "type": "string",
                                        "example": "7480201"
                                    },
                                    "subject_groups": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "A00",
                                            "A01"
                                        ]
                                    },
                                    "quota": {
                                        "type": "integer",
                                        "example": 120
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tao thanh cong"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admission-majors/{admission_major}": {
            "get": {
                "tags": [
                    "Admission Majors"
                ],
                "summary": "Chi tiet nganh tuyen sinh",
                "operationId": "7707a0e9cbb4967e378fff15ed183ec6",
                "parameters": [
                    {
                        "name": "admission_major",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chi tiet nganh"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admission Majors"
                ],
                "summary": "Cap nhat nganh tuyen sinh",
                "operationId": "a71517f08e65bb0665821a9960df5dc6",
                "parameters": [
                    {
                        "name": "admission_major",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "major_name": {
                                        "type": "string",
                                        "example": "Cong nghe thong tin"
                                    },
                                    "quota": {
                                        "type": "integer",
                                        "example": 150
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cap nhat thanh cong"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admission Majors"
                ],
                "summary": "Xoa nganh tuyen sinh",
                "operationId": "d53886b58453d510ee6a72e351cd7dec",
                "parameters": [
                    {
                        "name": "admission_major",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xoa thanh cong"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Dang nhap admin",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "admin@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "admin123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Dang nhap thanh cong"
                    },
                    "422": {
                        "description": "Sai thong tin dang nhap"
                    }
                }
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Lay thong tin admin hien tai",
                "operationId": "8367feff40ba10e43187792cad3132c1",
                "responses": {
                    "200": {
                        "description": "Thong tin user"
                    },
                    "401": {
                        "description": "Chua dang nhap"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Dang xuat admin",
                "operationId": "ad65cbbd4e9f201619eae184a5961a98",
                "responses": {
                    "200": {
                        "description": "Dang xuat thanh cong"
                    },
                    "401": {
                        "description": "Chua dang nhap"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Tao tai khoan admin",
                "operationId": "8e4e3cb7b46126f4696379d3c7eeb4ad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Admin 2"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "admin2@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "admin123456"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "admin123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tao tai khoan thanh cong"
                    },
                    "422": {
                        "description": "Du lieu khong hop le"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Nhập theo định dạng: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}